+++ Classes/Public/Manager/Submanagers/OCTSubmanagerChats.h
@@ -35,13 +35,14 @@
 - (void)removeMessages:(NSArray<OCTMessageAbstract *> *)messages;
 
 /**
- * Removes chat and all appropriate messages.
+ * Removes all messages in chat and chat itself.
  *
- * @param chat Chat to remove.
+ * @param chat Chat to remove in.
+ * @param removeChat Whether remove chat or not
  *
  * @warning Destructive operation! There is no way to restore chat or messages after removal.
  */
-- (void)removeChatWithAllMessages:(OCTChat *)chat;
+- (void)removeAllMessagesInChat:(OCTChat *)chat removeChat:(BOOL)removeChat;
 
 /**
  * Send text message to specific chat